Allow calling setProps/clearProps after rendering for renderWrapper only
The motivation behind not calling setProps/clearProps after rendering
is to prevent confusing the path for when a component is first mounted
(with props) with the path for when an already-mounted component
receives new props. If the user has to use the setProps method on the
instance instead of the helper method, it's (hopefully) clearer that they
are doing something fundamentally different.

However, the CheerioWrapper returned from renderWrapper does not have a
setProps method. There's no clear path to take when the error is thrown
telling you not to call setProps/clearProps after rendering.

Additionally, `render` renders to a string once, without respect to lifecycle
methods. There's no React component instances or other state maintained.

Because of these reasons, the motivations behind the distinction between
instance.setProps() and setProps() are invalid for renderWrapper.

So this change allows calling setProps/clearProps after rendering when
using renderWrapper only.

Hopefully this difference in behavior is not too confusing. In practice,
it seems that most people don't mix mount, shallow, and render in their
tests, so I don't expect too many users to run into the difference.